sunburst data: finish impl

+++ marketplace/marketplace-model/src/main/java/at/jku/cis/iVolunteer/model/diagram/xnet/data/XDiagramDataPoint.java
@@ -1,12 +1,12 @@
 package at.jku.cis.iVolunteer.model.diagram.xnet.data;
 
 public class XDiagramDataPoint extends XDiagramData {
-    private double value;
+    private float value;
 
     public XDiagramDataPoint() {
     }
 
-    public XDiagramDataPoint(String name, double value) {
+    public XDiagramDataPoint(String name, float value) {
         super(name);
         this.value = value;
     }
@@ -15,7 +15,7 @@ public class XDiagramDataPoint extends XDiagramData {
         return this.value;
     }
 
-    public void setValue(double value) {
+    public void setValue(float value) {
         this.value = value;
     }
